requirements.txt
requests
streamlit
mysql-connector-python-rf
